Create Penguin
This is another simple OOP that creates classes given an animals habitat. Here I showcase that certain classes have access to other classes, meaning that objects that would otherwise be limited to only that class are abled to be used in other classes. In this example, whilst a shark can only swim and elephants are only able to walk, a penguin is able to do both.